Rename & Remove Comments
