+++ src/pages/Auth/Register/CreateAccount/CreateAccount.module.scss
@@ -1,4 +1,4 @@
-@use "../../../variables";
+@use "../../../../variables";
 
 .register_page {
 	padding-top: 3rem;
