+++ src/pages/Auth/Register/CreateUserProfile/CreateUserProfile.tsx
@@ -1,0 +1,9 @@
+const CreateUserProfile = () => {
+	return (
+		<>
+			<h1>create user info</h1>
+		</>
+	);
+};
+
+export default CreateUserProfile;
